refactor(ui): move machine storage utils to node storage utils
