+++ ui/src/app/machines/components/MachineHeaderForms/ActionFormWrapper/CloneForm/CloneFormFields/CloneStorageTable/CloneStorageTable.tsx
@@ -13,7 +13,7 @@ import {
   formatSize,
   formatType,
   partitionAvailable,
-} from "app/store/machine/utils";
+} from "app/store/utils";
 
 type Props = {
   loadingMachineDetails?: boolean;
