+++ ui/src/app/machines/views/MachineDetails/MachineStorage/AvailableStorageTable/BulkActions/BulkActions.tsx
@@ -10,16 +10,16 @@ import UpdateDatastore from "./UpdateDatastore";
 
 import machineSelectors from "app/store/machine/selectors";
 import type { Machine } from "app/store/machine/types";
+import { isMachineDetails } from "app/store/machine/utils";
+import type { RootState } from "app/store/root/types";
+import { StorageLayout } from "app/store/types/enum";
+import type { Disk, Partition } from "app/store/types/node";
 import {
   canCreateOrUpdateDatastore,
   canCreateRaid,
   canCreateVolumeGroup,
   isDatastore,
-  isMachineDetails,
-} from "app/store/machine/utils";
-import type { RootState } from "app/store/root/types";
-import { StorageLayout } from "app/store/types/enum";
-import type { Disk, Partition } from "app/store/types/node";
+} from "app/store/utils";
 
 type Props = {
   bulkAction: BulkAction | null;
