+++ ui/src/app/machines/views/MachineDetails/MachineStorage/AvailableStorageTable/BulkActions/CreateDatastore/CreateDatastore.tsx
@@ -18,15 +18,15 @@ import { actions as machineActions } from "app/store/machine";
 import machineSelectors from "app/store/machine/selectors";
 import type { Machine } from "app/store/machine/types";
 import type { MachineEventErrors } from "app/store/machine/types/base";
+import { isMachineDetails } from "app/store/machine/utils";
+import type { RootState } from "app/store/root/types";
+import type { Disk, Partition } from "app/store/types/node";
 import {
   formatSize,
   formatType,
   isDatastore,
-  isMachineDetails,
   splitDiskPartitionIds,
-} from "app/store/machine/utils";
-import type { RootState } from "app/store/root/types";
-import type { Disk, Partition } from "app/store/types/node";
+} from "app/store/utils";
 
 type CreateDatastoreValues = {
   name: string;
