+++ ui/src/app/machines/views/MachineDetails/MachineStorage/AvailableStorageTable/BulkActions/UpdateDatastore/UpdateDatastoreFields/UpdateDatastoreFields.tsx
@@ -13,8 +13,8 @@ import { useFormikContext } from "formik";
 import type { UpdateDatastoreValues } from "../UpdateDatastore";
 
 import FormikField from "app/base/components/FormikField";
-import { formatSize, formatType } from "app/store/machine/utils";
 import type { Disk, Partition } from "app/store/types/node";
+import { formatSize, formatType } from "app/store/utils";
 
 type Props = {
   datastores: Disk[];
