+++ ui/src/app/machines/views/MachineDetails/MachineStorage/AvailableStorageTable/EditPartition/EditPartition.test.tsx
@@ -6,12 +6,12 @@ import EditPartition from "./EditPartition";
 
 import {
   machineDetails as machineDetailsFactory,
-  machineDisk as diskFactory,
   machineEventError as machineEventErrorFactory,
-  machinePartition as partitionFactory,
   machineState as machineStateFactory,
   machineStatus as machineStatusFactory,
   machineStatuses as machineStatusesFactory,
+  nodeDisk as diskFactory,
+  nodePartition as partitionFactory,
   rootState as rootStateFactory,
 } from "testing/factories";
 import { submitFormikForm } from "testing/utils";
