+++ ui/src/app/machines/views/MachineDetails/MachineStorage/AvailableStorageTable/FilesystemFields/FilesystemFields.tsx
@@ -5,9 +5,10 @@ import { useSelector } from "react-redux";
 import FormikField from "app/base/components/FormikField";
 import machineSelectors from "app/store/machine/selectors";
 import type { Machine } from "app/store/machine/types";
-import { isMachineDetails, usesStorage } from "app/store/machine/utils";
+import { isMachineDetails } from "app/store/machine/utils";
 import type { RootState } from "app/store/root/types";
 import type { Filesystem } from "app/store/types/node";
+import { usesStorage } from "app/store/utils";
 
 type FilesystemValues = {
   fstype: Filesystem["fstype"];
