+++ ui/src/app/machines/views/MachineDetails/MachineStorage/ChangeStorageLayout/ChangeStorageLayout.tsx
@@ -10,8 +10,8 @@ import { useMachineDetailsForm } from "app/machines/hooks";
 import { actions as machineActions } from "app/store/machine";
 import type { Machine } from "app/store/machine/types";
 import type { MachineEventErrors } from "app/store/machine/types/base";
-import { isVMWareLayout } from "app/store/machine/utils";
 import { StorageLayout } from "app/store/types/enum";
+import { isVMWareLayout } from "app/store/utils";
 
 type StorageLayoutOption = {
   label: string;
