+++ ui/src/app/machines/views/MachineDetails/MachineStorage/DatastoresTable/DatastoresTable.tsx
@@ -10,12 +10,9 @@ import TableActionsDropdown from "app/base/components/TableActionsDropdown";
 import { actions as machineActions } from "app/store/machine";
 import machineSelectors from "app/store/machine/selectors";
 import type { Machine } from "app/store/machine/types";
-import {
-  formatSize,
-  isDatastore,
-  isMachineDetails,
-} from "app/store/machine/utils";
+import { isMachineDetails } from "app/store/machine/utils";
 import type { RootState } from "app/store/root/types";
+import { formatSize, isDatastore } from "app/store/utils";
 
 export enum DatastoreAction {
   DELETE = "deleteDisk",
