+++ ui/src/app/machines/views/MachineDetails/MachineStorage/FilesystemsTable/FilesystemsTable.tsx
@@ -12,14 +12,10 @@ import TableActionsDropdown from "app/base/components/TableActionsDropdown";
 import { actions as machineActions } from "app/store/machine";
 import machineSelectors from "app/store/machine/selectors";
 import type { Machine } from "app/store/machine/types";
-import {
-  formatSize,
-  isMachineDetails,
-  isMounted,
-  usesStorage,
-} from "app/store/machine/utils";
+import { isMachineDetails } from "app/store/machine/utils";
 import type { RootState } from "app/store/root/types";
 import type { Filesystem, Disk, Partition } from "app/store/types/node";
+import { formatSize, isMounted, usesStorage } from "app/store/utils";
 
 export enum FilesystemAction {
   DELETE = "deleteFilesystem",
