+++ ui/src/app/machines/views/MachineDetails/MachineStorage/StorageNotifications/StorageNotifications.test.tsx
@@ -10,8 +10,8 @@ import type { RootState } from "app/store/root/types";
 import { NodeStatusCode } from "app/store/types/node";
 import {
   machineDetails as machineDetailsFactory,
-  machineDisk as diskFactory,
   machineState as machineStateFactory,
+  nodeDisk as diskFactory,
   rootState as rootStateFactory,
 } from "testing/factories";
