+++ ui/src/app/settings/views/Storage/StorageForm/StorageFormFields/StorageFormFields.tsx
@@ -6,8 +6,8 @@ import type { StorageFormValues } from "../types";
 
 import FormikField from "app/base/components/FormikField";
 import configSelectors from "app/store/config/selectors";
-import { isVMWareLayout } from "app/store/machine/utils";
 import { StorageLayout } from "app/store/types/enum";
+import { isVMWareLayout } from "app/store/utils";
 
 const StorageFormFields = (): JSX.Element => {
   const { values } = useFormikContext<StorageFormValues>();
