+++ ui/src/app/store/machine/utils/index.ts
@@ -22,40 +22,3 @@ export {
 } from "./search";
 
 export { isTransientStatus } from "./status";
-
-export {
-  canBeDeleted,
-  canBeFormatted,
-  canBePartitioned,
-  canCreateBcache,
-  canCreateCacheSet,
-  canCreateLogicalVolume,
-  canCreateOrUpdateDatastore,
-  canCreateRaid,
-  canCreateVolumeGroup,
-  canOsSupportBcacheZFS,
-  canOsSupportStorageConfig,
-  canSetBootDisk,
-  diskAvailable,
-  formatSize,
-  formatType,
-  getDiskById,
-  getNextStorageName,
-  getPartitionById,
-  isBcache,
-  isCacheSet,
-  isDatastore,
-  isDisk,
-  isLogicalVolume,
-  isMachineStorageConfigurable,
-  isMounted,
-  isPartition,
-  isPhysical,
-  isRaid,
-  isVirtual,
-  isVMWareLayout,
-  isVolumeGroup,
-  partitionAvailable,
-  splitDiskPartitionIds,
-  usesStorage,
-} from "./storage";
