+++ ui/src/app/store/utils/index.ts
@@ -37,6 +37,42 @@ export {
   isInterfaceConnected,
   LINK_MODE_DISPLAY,
 } from "./node/networking";
+export {
+  canBeDeleted,
+  canBeFormatted,
+  canBePartitioned,
+  canCreateBcache,
+  canCreateCacheSet,
+  canCreateLogicalVolume,
+  canCreateOrUpdateDatastore,
+  canCreateRaid,
+  canCreateVolumeGroup,
+  canOsSupportBcacheZFS,
+  canOsSupportStorageConfig,
+  canSetBootDisk,
+  diskAvailable,
+  formatSize,
+  formatType,
+  getDiskById,
+  getNextStorageName,
+  getPartitionById,
+  isBcache,
+  isCacheSet,
+  isDatastore,
+  isDisk,
+  isLogicalVolume,
+  isMounted,
+  isNodeStorageConfigurable,
+  isPartition,
+  isPhysical,
+  isRaid,
+  isVirtual,
+  isVMWareLayout,
+  isVolumeGroup,
+  partitionAvailable,
+  splitDiskPartitionIds,
+  usesStorage,
+} from "./node/storage";
 export { generateBaseSelectors } from "./selectors";
 export {
   generateCommonReducers,
