+++ ui/src/testing/factories/index.ts
@@ -102,17 +102,17 @@ export {
   machine,
   machineDetails,
   machineDevice,
-  machineDisk,
   machineEvent,
   machineEventType,
-  machineFilesystem,
   machineInterface,
   machineIpAddress,
   machineNumaNode,
-  machinePartition,
   networkDiscoveredIP,
   networkInterface,
   networkLink,
+  nodeDisk,
+  nodeFilesystem,
+  nodePartition,
   pod,
   podDetails,
   podMemoryResource,
